Adapt to API changes in Spring AI 1.0.0-M6.

+++ observability-example/src/main/java/com/packt/spring/ai/examples/observability/TokenCountObservationApplication.java
@@ -102,7 +102,7 @@ public class TokenCountObservationApplication extends AbstractSpringBootApplicat
 
 		ChatResponseMetadata metadata = chatResponse.getMetadata();
 
-		return metadata.getUsage().getTotalTokens();
+		return metadata.getUsage().getTotalTokens().longValue();
 	}
 
 	@SuppressWarnings("all")
